refactored Storable service
