+++ broker/core/src/test/resources/locator.xml
@@ -20,9 +20,6 @@
         <api>org.eclipse.kapua.service.certificate.info.CertificateInfoService</api>
         <api>org.eclipse.kapua.service.certificate.info.CertificateInfoFactory</api>
 
-        <api>org.eclipse.kapua.service.storable.model.id.StorableIdFactory</api>
-        <api>org.eclipse.kapua.service.storable.model.query.predicate.StorablePredicateFactory</api>
-
         <api>org.eclipse.kapua.message.KapuaMessageFactory</api>
 
         <api>org.eclipse.kapua.model.id.KapuaIdFactory</api>
