+++ qa/integration/src/test/resources/locator.xml
@@ -21,9 +21,6 @@
 
         <api>org.eclipse.kapua.message.device.lifecycle.KapuaLifecycleMessageFactory</api>
 
-        <api>org.eclipse.kapua.service.storable.model.id.StorableIdFactory</api>
-        <api>org.eclipse.kapua.service.storable.model.query.predicate.StorablePredicateFactory</api>
-
         <api>org.eclipse.kapua.job.engine.JobEngineService</api>
         <api>org.eclipse.kapua.job.engine.JobEngineFactory</api>
