+++ service/commons/storable/internal/src/main/java/org/eclipse/kapua/service/storable/model/id/StorableIdModule.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2016, 2021 Eurotech and/or its affiliates and others
+ * Copyright (c) 2021 Eurotech and/or its affiliates and others
  *
  * This program and the accompanying materials are made
  * available under the terms of the Eclipse Public License 2.0
@@ -12,18 +12,11 @@
  *******************************************************************************/
 package org.eclipse.kapua.service.storable.model.id;
 
-import javax.inject.Singleton;
+import org.eclipse.kapua.commons.core.AbstractKapuaModule;
 
-/**
- * {@link StorableIdFactory} implementation.
- *
- * @since 1.0.0
- */
-@Singleton
-public class StorableIdFactoryImpl implements StorableIdFactory {
-
+public class StorableIdModule extends AbstractKapuaModule {
     @Override
-    public StorableId newStorableId(String stringId) {
-        return new StorableIdImpl(stringId);
+    protected void configureModule() {
+        bind(StorableIdFactory.class).to(StorableIdFactoryImpl.class);
     }
 }
